pyproject.toml
[tool.poetry]
name = "raglaba6"
version = "0.1.0"
description = "RAG service for 10-K financial reports with clean architecture"
authors = ["user <user@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = ">=0.115.0"
uvicorn = {version = ">=0.32.0", extras = ["standard"]}
qdrant-client = ">=1.12.0"
sentence-transformers = ">=3.0.0"
langchain-text-splitters = ">=0.3.0"
docling = ">=2.0.0"
docling-surya = {version = ">=0.1.0", python = ">=3.12,<3.14"}
python-dotenv = ">=1.0.0"
requests = ">=2.32.0"
pydantic = ">=2.9.0"
python-multipart = ">=0.0.12"
punq = ">=0.7.0"
pypdf = "^6.10.2"
transformers = "<5"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
